# platform push: Avoid fetching image repo metadata when not needed