The platform push command was always fetching the image repo metadata even when that data was not used. Now we only fetch the data when the switch --compatible-with is passed; this makes the command run faster when the switch is not passed.

Here's an example of running the recently added command to push fuse packages before and after this improvement:

```
# Before:
$ time torizoncore-builder tcb485 platform push ./fuses-missing.yaml --hardwareid colibri-imx8x-fuses --credentials ./credentials-prod.zip
Error: Found 9 fuse values in provided yaml file, but expected to find 16 instead.
exit status 255

real    0m18,253s <=
user    0m0,011s
sys     0m0,008s

# After:
$ time torizoncore-builder platform push ./fuses-missing.yaml --hardwareid colibri-imx8x-fuses --credentials ./credentials-prod.zip
Error: Found 9 fuse values in provided yaml file, but expected to find 16 instead.
exit status 255

real    0m0,557s <=
user    0m0,013s
sys     0m0,024s
```